[FLINK-36316][API / Core] Move PortRange class from flink-runtime to flink-core module
