+++ flink-core/src/test/java/org/apache/flink/util/PortRangeTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.runtime.util;
+package org.apache.flink.util;
 
 import org.apache.flink.configuration.IllegalConfigurationException;
